Eliminar ventas

+++ flaskProject/app.py
@@ -15,7 +15,7 @@ import json
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = 'dev'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://lab:Developer123!@localhost:3306/base_merkaap'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://ing:Developer123!@localhost:3306/base_merkaap'
 db.init_app(app)
 CORS(app)
 
@@ -60,17 +60,11 @@ def login():
 def recuperarVentas(): 
     if request.method == 'OPTIONS':                   
         res = Response()        
-        print()
-        print(res)
-        print()
         res.headers['X-Content-Type-Options'] = '*'
         return res
-    elif request.method == 'POST':
-        print()
+    elif request.method == 'POST':   
         print(request.json)
-        if request.json == None:
-            print("ERROR")
-        print()
+        #exit(0)   
         correo = request.json[2]
         contrasenia = request.json[3]
         vendedor = Vendedor.query.filter(Vendedor.correo == correo, Vendedor.contrasenia == contrasenia).first()
@@ -153,20 +147,22 @@ def eliminarVentas():
         return res
     elif request.method == 'GET':
         return json.dumps({'hola': 'fin'})
-    elif request.method == 'POST':
-        lista_de_id = request.json
+    elif request.method == 'POST':        
+        lista_de_id = request.json      
+        print(lista_de_id)  
         for id in lista_de_id:
             venta = Producto.query.filter(Producto.idProducto == id).first()
-            for registro in Comprar.query.filter(Comprar.idProducto == id).all():
-                db.session.delete(registro)
-                db.session.commit()
+            if(Comprar.query.filter(Comprar.idProducto == id).first()):
+                for registro in Comprar.query.filter(Comprar.idProducto == id).all():
+                    db.session.delete(registro)
+                    db.session.commit()
             db.session.delete(venta)
             db.session.commit()
-        id = 1
+        id = 1        
         for registro in Producto.query.all():
-            registro.idProducto = id
+            registro.idProducto = id                        
             db.session.commit()
-            id += 1
+            id += 1        
         return json.dumps({'listo':'usuario'})
 
 @app.route('/recuperarProductos', methods=['GET', 'POST', 'OPTIONS'])
@@ -192,12 +188,13 @@ def recuperarProducto():
     elif request.method == 'POST':
         id = request.json
         producto = Producto.query.filter(Producto.idProducto == id).first()                
-        categoria = Categoria.query.filter(Categoria.idCategoria == producto.idCategoria).first()        
-        comentarios = []
+        categoria = Categoria.query.filter(Categoria.idCategoria == producto.idCategoria).first()
+        vendedor = Vendedor.query.filter(Vendedor.idVendedor == producto.idVendedor).first()        
+        comentarios = []        
         for comentario in Comprar.query.filter(Comprar.idProducto == producto.idProducto).all():
             comprador = Comprador.query.filter(Comprador.idComprador == comentario.idComprador).first()
-            comentarios.append([comentario.calificacion, comprador.nombre, comentario.comentario])
-        producto = {producto.idProducto:[producto.nombre, categoria.nombre, producto.precio , producto.stock, comentarios]}
+            comentarios.append([comentario.calificacion, comprador.nombre, comentario.comentario])                    
+        producto = {producto.idProducto:[producto.nombre, categoria.nombre, producto.precio , producto.stock, comentarios, vendedor.nombre, vendedor.apellidoPat, vendedor.correo, vendedor.telefono]}
         return json.dumps({'producto': producto})
 
 @app.route('/agregarOpinion', methods=['GET', 'POST', 'OPTIONS'])
@@ -228,8 +225,48 @@ def agregarOpinion():
             compra = Comprar(id, comprador.idComprador, int(idProducto), opinion[1], opinion[0])
             db.session.add(compra)
             db.session.commit()      
-        return json.dumps({'listo':'usuario'})
+        return json.dumps({'Op':[opinion[0],opinion]})
     
+@app.route('/comprarProducto', methods=['GET', 'POST', 'OPTIONS'])
+def comprarProducto():
+    if request.method == 'OPTIONS':
+        res = Response()        
+        return res
+    elif request.method == 'POST':                 
+        id = request.json[0]
+        usuario = request.json[1]        
+        correo = usuario[0][1][2]                 
+        producto = Producto.query.filter(Producto.idProducto == id).first()
+        vendedor = Vendedor.query.filter(Vendedor.idVendedor == producto.idVendedor).first()
+        producto.stock = producto.stock-1                
+        if producto.stock >= 0:
+            db.session.commit()  
+            # Información del correo
+            correo_merkapp = "merkapp.online@gmail.com"
+            # Debería ser una contraseña de aplicación o una contraseña segura
+            contra_merkapp = "edlu ylzl vbrw ugzf"  
+
+            msg = EmailMessage()
+            msg["Subject"] = "¡Compra exitosa!"
+            msg["From"] = correo_merkapp
+            msg["To"] = correo
+            msg.set_content(f"Estimado(a) {usuario[0][1][1]},\n\nNos complace informarte que tu compra en MerkApp ha sido procesada con éxito.\nDetalles de la compra:\n - Producto: {producto.nombre}\n - Total: ${producto.precio}.00\n\nEn las próximas horas, el vendedor {vendedor.nombre} se pondrá en contacto contigo para coordinar los detalles de la entrega de tu pedido.\nRecuerda que puedes comunicarte directamente con el vendedor a través de correo electrónico = {vendedor.correo} o teléfono = {vendedor.telefono} para cualquier duda o consulta sobre tu compra.\n\nAgradecemos tu confianza en MerkApp y esperamos que disfrutes de tu nueva adquisición.\n\nAtentamente, MerkApp")
+
+            try:
+                # Usar SMTP_SSL si el puerto es 465
+                with smtplib.SMTP_SSL("smtp.gmail.com", port=465) as smtp:
+                    smtp.login(correo_merkapp, contra_merkapp)
+                    print(f"Enviando correo")
+                    smtp.send_message(msg)
+                    print(f"Correo enviado")
+            except smtplib.SMTPException as e:
+                return json.dumps({'error': 'No se envió correo'})
+            if producto.stock == 0:
+                db.session.delete(producto)
+                db.session.commit()
+            return json.dumps({'listo':'usuario'})
+        else:
+            return json.dumps({'error': 'Inventario vació'})           
 
 @app.route('/buscar', methods=['GET'])
 def buscar_productos():
@@ -316,11 +353,13 @@ def register():
                 smtp.send_message(msg)
                 print(f"Correo enviado")
         except smtplib.SMTPException as e:
-            print(f"Error al enviar el correo a {correo}: {e}")
-            
+            print(f"Error al enviar el correo a {correo}: {e}")                    
+
         db.session.add(nuevo_usuario)
         db.session.commit()
         return json.dumps({'listo':nombre, 'correo':correo})
+
+
 
 if __name__ == '__main__':
     app.run()
